Langchain1/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
chromadb==0.3.25
openai==1.8.6
python-dotenv==1.0.0
requests==2.31.0
pydantic==1.10.12
